+++ app/src/app/regionen/[regionSlug]/_components/SidebarInspector/Tools/ToolsOtherProperties.tsx
@@ -1,7 +1,11 @@
 import { Link } from '@/src/app/_components/links/Link'
+import { LinkExternal } from '@/src/app/_components/links/LinkExternal'
+import { isProd } from '@/src/app/_components/utils/isEnv'
 import { format, formatDistanceToNow, fromUnixTime } from 'date-fns'
 import { de } from 'date-fns/locale'
+import { useMapState } from '../../../_hooks/mapState/useMapState'
 import { InspectorFeature } from '../Inspector'
+import { tilesInspectorWithGeomUrl } from './osmUrls/osmUrls'
 
 type Props = {
   feature: InspectorFeature['feature']
@@ -9,6 +13,11 @@ type Props = {
 }
 
 export const ToolsOtherProperties = ({ feature, documentedKeys }: Props) => {
+  const inspectorOtherPropertiesOpen = useMapState((state) => state.inspectorOtherPropertiesOpen)
+  const setInspectorOtherPropertiesOpen = useMapState(
+    (state) => state.actions.setInspectorOtherPropertiesOpen,
+  )
+
   const systemKeys = [
     '_todos',
     'fresh',
@@ -40,8 +49,20 @@ export const ToolsOtherProperties = ({ feature, documentedKeys }: Props) => {
       ([key, _v]) => systemKeys.includes(key) && documentedKeys && !documentedKeys?.includes(key),
     )
 
+  const viewerUrl =
+    !isProd && feature.sourceLayer && feature.geometry
+      ? tilesInspectorWithGeomUrl({
+          geometry: feature.geometry,
+          sourceLayer: feature.sourceLayer,
+        })
+      : undefined
+
   return (
-    <details className="mt-3">
+    <details
+      className="mt-3"
+      open={inspectorOtherPropertiesOpen}
+      onToggle={(e) => setInspectorOtherPropertiesOpen(e.currentTarget.open)}
+    >
       <summary className="ml-1.5 cursor-pointer font-semibold text-gray-600">
         <span className="ml-1.5">Weitere Daten an diesem Element</span>
       </summary>
@@ -79,6 +100,20 @@ export const ToolsOtherProperties = ({ feature, documentedKeys }: Props) => {
               <code>feature.id</code>
             </strong>
             : {feature.id || 'MISSING'}
+          </p>
+          <p className="mb-0.5 border-b border-gray-200 pb-0.5">
+            <strong>
+              <code>sourceLayer</code>
+            </strong>
+            : {feature.sourceLayer || 'UNBEKANNT'}
+            {viewerUrl && (
+              <>
+                {' '}
+                <LinkExternal blank href={viewerUrl} className="scale-75">
+                  Viewer
+                </LinkExternal>
+              </>
+            )}
           </p>
           {systemProperties.length ? (
             systemProperties.map(([key, value]) => {