+++ app/src/app/regionen/[regionSlug]/_mapData/mapDataSubcategories/defaultStyle/defaultStyleHidden.ts
@@ -3,7 +3,6 @@ import { FileMapDataSubcategoryHiddenStyle } from '../../types'
 export const defaultStyleHidden: FileMapDataSubcategoryHiddenStyle = {
   id: 'hidden',
   name: 'Kein Stil',
-  desc: null,
   layers: undefined,
   legends: undefined,
 }