+++ processing/topics/parking/obstacles/0_areas_project_to_kerb.sql
@@ -96,19 +96,22 @@ FROM
   _parking_obstacle_areas a
   JOIN _parking_separate_parking_areas_projected spap ON a.separate_parking_area_id = spap.source_id;
 
--- Filter: remove furthest projection if multiple
+-- Filter: remove duplicate projections to the same kerb
 -- `tilda_project_to_k_closest_kerbs` can return up to 6 projections (k := 6) for each obstacle area.
--- When an obstacle area has multiple projections, keep only the closest ones (remove furthest).
+-- When an obstacle area has multiple projections to the same kerb, keep only the closest one (remove furthest).
+-- For intersection restricted areas, we want to keep projections to different kerbs (different streets).
 DELETE FROM _parking_obstacle_areas_projected
 WHERE
-  (source_id, kerb_distance) IN (
+  (source_id, kerb_id, kerb_distance) IN (
     SELECT
       source_id,
+      kerb_id,
       MAX(kerb_distance)
     FROM
       _parking_obstacle_areas_projected
     GROUP BY
-      source_id
+      source_id,
+      kerb_id
     HAVING
       count(*) > 1
   );